Updating challenge

+++ routes/mission.js
@@ -8,42 +8,45 @@ const {mustBeAdmin, mustBeLoggedIn, selfOnly} = require('./permissions')
 
 router.post('/', function(req, res, next){
 	console.log("posting mission")
-	console.log("REQ BODY FROM FORM: ", req.body)
-	//mustBeAdmin()(req, res, next)
-	let {title, description, place, meetingPlace, location} = req.body
-	Mission.create({
-		title, description, place, meetingPlace, location, 
-		numChallenges: 0
-	})
-	.then(mission => {
-		res.status(200).json(mission)
-	})
-	.catch(next)
+	if(mustBeAdmin()(req, res, next) === "continue"){
+		let {title, description, place, meetingPlace, location} = req.body
+		Mission.create({
+			title, description, place, meetingPlace, location, 
+			numChallenges: 0
+		})
+		.then(mission => {
+			res.status(200).json(mission)
+		})
+		.catch(next)
+	}
 })
 
 
 router.put('/:id/update', function(req, res, next){
 	let {title, description, place, meetingPlace, location} = req.body
-	Mission.findById(req.params.id)
-	.then(mission => {
-		mission.update({
-			title, description, place, meetingPlace, location
+	if(mustBeAdmin()(req, res, next) === "continue"){
+		Mission.findById(req.params.id)
+		.then(mission => {
+			mission.update({
+				title, description, place, meetingPlace, location
+			})
+			.then(mission => res.status(200).json(mission))
 		})
-		.then(mission => res.status(200).json(mission))
-	})
+	}
 })
 
 router.delete('/:id', function(req, res, next){
-	//mustBeAdmin()(req, res, next)
-	Mission.findById(req.params.id)
-	.then((mission) => {
-		mission.getChallenges()
-		.then(() => {
-			console.log("DON'T FORGET TO UPDATE CHALLENGES")
-			return mission.destroy({force: true})
+	if(mustBeAdmin()(req, res, next) === "continue"){
+		Mission.findById(req.params.id)
+		.then((mission) => {
+			mission.getChallenges()
+			.then(() => {
+				console.log("DON'T FORGET TO UPDATE CHALLENGES")
+				return mission.destroy({force: true})
+			})
+			.then(() => res.sendStatus(200))
 		})
-		.then(() => res.sendStatus(200))
-	})
+	}
 })
 
 module.exports = router;
